changed fabs() to fabsf()

+++ assign 2/assn2.c
@@ -325,7 +325,7 @@ void rgb_to_hsv(int image_rgb[][SIZE][SIZE], float image_hsv[][SIZE][SIZE], int 
 void dot_hsv_to_rgb(float h, float s, float v, int *r, int *g, int *b) {
     /* 모든 변수의 이름은 pdf instruction을 따릅니다. */
     float c = v * s; // v와 s를 곱해 c를 구한다. 
-    float x = c * (1 - fabs(fmod(h / 60.0, 2) - 1)); // 공식을 이용해 x를 구한다. 
+    float x = c * (1 - fabsf(fmod(h / 60.0, 2) - 1)); // 공식을 이용해 x를 구한다. 
     float m = v - c; // v와 c의 차이로 m을 구한다. 
 
     float _r, _g, _b; // normalized r, g, b를 임시 저장할 변수를 선언한다.
